Create coke.py
A program that prompts the user to insert a coin, one at a time, each time informing the user of the amount due. Once the user has inputted at least 50 cents, it outputs how many cents in change the user is owed. It assumes that the user will only input integers, and ignores any integer that isn’t an accepted denomination.

The only accepted denominations are: 5, 10 and 25 cents.